Move CodecHandlers into separate source/headers
