+++ src/codechandler/MPEGCodecHandler.h
@@ -1,0 +1,7 @@
+#include "CodecHandler.h"
+
+class ATTR_DLL_LOCAL MPEGCodecHandler : public CodecHandler
+{
+public:
+  MPEGCodecHandler(AP4_SampleDescription* sd);
+};
