+++ src/codechandler/VP9CodecHandler.h
@@ -1,0 +1,7 @@
+#include "CodecHandler.h"
+
+class ATTR_DLL_LOCAL VP9CodecHandler : public CodecHandler
+{
+public:
+  VP9CodecHandler(AP4_SampleDescription* sd);
+};
