+++ src/main.cpp
@@ -16,10 +16,15 @@
 #include "helpers.h"
 #include "log.h"
 #include "oscompat.h"
+#include "codechandler/AVCCodecHandler.h"
+#include "codechandler/HEVCCodecHandler.h"
+#include "codechandler/MPEGCodecHandler.h"
+#include "codechandler/TTMLCodecHandler.h"
+#include "codechandler/VP9CodecHandler.h"
+#include "codechandler/WebVTTCodecHandler.h"
 #include "parser/DASHTree.h"
 #include "parser/HLSTree.h"
 #include "parser/SmoothTree.h"
-#include "parser/TTML.h"
 
 #include <chrono>
 #include <algorithm>
@@ -515,520 +520,6 @@ bool KodiAdaptiveStream::parseIndexRange(adaptive::AdaptiveTree::Representation*
   }
   return false;
 }
-
-/*******************************************************
-|   CodecHandler
-********************************************************/
-
-class ATTR_DLL_LOCAL CodecHandler
-{
-public:
-  CodecHandler(AP4_SampleDescription* sd)
-    : sample_description(sd), naluLengthSize(0), pictureId(0), pictureIdPrev(0xFF){};
-  virtual ~CodecHandler(){};
-
-  virtual void UpdatePPSId(AP4_DataBuffer const&){};
-  virtual bool GetInformation(kodi::addon::InputstreamInfo& info)
-  {
-    AP4_GenericAudioSampleDescription* asd(nullptr);
-    if (sample_description)
-    {
-      if ((asd = dynamic_cast<AP4_GenericAudioSampleDescription*>(sample_description)))
-      {
-        if ((!info.GetChannels() && asd->GetChannelCount() != info.GetChannels()) ||
-            (!info.GetSampleRate() && asd->GetSampleRate() != info.GetSampleRate()) ||
-            (!info.GetBitsPerSample() && asd->GetSampleSize() != info.GetBitsPerSample()))
-        {
-          if (!info.GetChannels())
-            info.SetChannels(asd->GetChannelCount());
-          if (!info.GetSampleRate())
-            info.SetSampleRate(asd->GetSampleRate());
-          if (!info.GetBitsPerSample())
-            info.SetBitsPerSample(asd->GetSampleSize());
-          return true;
-        }
-      }
-      else
-      {
-        //Netflix Framerate
-        AP4_Atom* atom;
-        AP4_UnknownUuidAtom* nxfr;
-        static const AP4_UI08 uuid[16] = {0x4e, 0x65, 0x74, 0x66, 0x6c, 0x69, 0x78, 0x46,
-                                          0x72, 0x61, 0x6d, 0x65, 0x52, 0x61, 0x74, 0x65};
-
-        if ((atom = sample_description->GetDetails().GetChild(static_cast<const AP4_UI08*>(uuid),
-                                                              0)) &&
-            (nxfr = dynamic_cast<AP4_UnknownUuidAtom*>(atom)) &&
-            nxfr->GetData().GetDataSize() == 10)
-        {
-          AP4_UI16 fpsRate = nxfr->GetData().GetData()[7] | nxfr->GetData().GetData()[6] << 8;
-          AP4_UI16 fpsScale = nxfr->GetData().GetData()[9] | nxfr->GetData().GetData()[8] << 8;
-
-          if (info.GetFpsScale() != fpsScale || info.GetFpsRate() != fpsRate)
-          {
-            info.SetFpsScale(fpsScale);
-            info.SetFpsRate(fpsRate);
-            return true;
-          }
-        }
-      }
-    }
-    return false;
-  };
-  virtual bool ExtraDataToAnnexB() { return false; };
-  virtual STREAMCODEC_PROFILE GetProfile() { return STREAMCODEC_PROFILE::CodecProfileNotNeeded; };
-  virtual bool Transform(AP4_UI64 pts, AP4_UI32 duration, AP4_DataBuffer& buf, AP4_UI64 timescale)
-  {
-    return false;
-  };
-  virtual bool ReadNextSample(AP4_Sample& sample, AP4_DataBuffer& buf) { return false; };
-  virtual void SetPTSOffset(AP4_UI64 offset){};
-  virtual bool TimeSeek(AP4_UI64 seekPos) { return true; };
-  virtual void Reset(){};
-
-  AP4_SampleDescription* sample_description;
-  AP4_DataBuffer extra_data;
-  AP4_UI08 naluLengthSize;
-  AP4_UI08 pictureId, pictureIdPrev;
-};
-
-/***********************   AVC   ************************/
-
-class ATTR_DLL_LOCAL AVCCodecHandler : public CodecHandler
-{
-public:
-  AVCCodecHandler(AP4_SampleDescription* sd)
-    : CodecHandler(sd), countPictureSetIds(0), needSliceInfo(false)
-  {
-    unsigned int width(0), height(0);
-    if (AP4_VideoSampleDescription* video_sample_description =
-            AP4_DYNAMIC_CAST(AP4_VideoSampleDescription, sample_description))
-    {
-      width = video_sample_description->GetWidth();
-      height = video_sample_description->GetHeight();
-    }
-    if (AP4_AvcSampleDescription* avc =
-            AP4_DYNAMIC_CAST(AP4_AvcSampleDescription, sample_description))
-    {
-      extra_data.SetData(avc->GetRawBytes().GetData(), avc->GetRawBytes().GetDataSize());
-      countPictureSetIds = avc->GetPictureParameters().ItemCount();
-      naluLengthSize = avc->GetNaluLengthSize();
-      needSliceInfo = (countPictureSetIds > 1 || !width || !height);
-      switch (avc->GetProfile())
-      {
-        case AP4_AVC_PROFILE_BASELINE:
-          codecProfile = STREAMCODEC_PROFILE::H264CodecProfileBaseline;
-          break;
-        case AP4_AVC_PROFILE_MAIN:
-          codecProfile = STREAMCODEC_PROFILE::H264CodecProfileMain;
-          break;
-        case AP4_AVC_PROFILE_EXTENDED:
-          codecProfile = STREAMCODEC_PROFILE::H264CodecProfileExtended;
-          break;
-        case AP4_AVC_PROFILE_HIGH:
-          codecProfile = STREAMCODEC_PROFILE::H264CodecProfileHigh;
-          break;
-        case AP4_AVC_PROFILE_HIGH_10:
-          codecProfile = STREAMCODEC_PROFILE::H264CodecProfileHigh10;
-          break;
-        case AP4_AVC_PROFILE_HIGH_422:
-          codecProfile = STREAMCODEC_PROFILE::H264CodecProfileHigh422;
-          break;
-        case AP4_AVC_PROFILE_HIGH_444:
-          codecProfile = STREAMCODEC_PROFILE::H264CodecProfileHigh444Predictive;
-          break;
-        default:
-          codecProfile = STREAMCODEC_PROFILE::CodecProfileUnknown;
-          break;
-      }
-    }
-  }
-
-  virtual bool ExtraDataToAnnexB() override
-  {
-    if (AP4_AvcSampleDescription* avc =
-            AP4_DYNAMIC_CAST(AP4_AvcSampleDescription, sample_description))
-    {
-      //calculate the size for annexb
-      size_t sz(0);
-      AP4_Array<AP4_DataBuffer>& pps(avc->GetPictureParameters());
-      for (unsigned int i(0); i < pps.ItemCount(); ++i)
-        sz += 4 + pps[i].GetDataSize();
-      AP4_Array<AP4_DataBuffer>& sps(avc->GetSequenceParameters());
-      for (unsigned int i(0); i < sps.ItemCount(); ++i)
-        sz += 4 + sps[i].GetDataSize();
-
-      extra_data.SetDataSize(sz);
-      uint8_t* cursor(extra_data.UseData());
-
-      for (unsigned int i(0); i < sps.ItemCount(); ++i)
-      {
-        cursor[0] = cursor[1] = cursor[2] = 0;
-        cursor[3] = 1;
-        memcpy(cursor + 4, sps[i].GetData(), sps[i].GetDataSize());
-        cursor += sps[i].GetDataSize() + 4;
-      }
-      for (unsigned int i(0); i < pps.ItemCount(); ++i)
-      {
-        cursor[0] = cursor[1] = cursor[2] = 0;
-        cursor[3] = 1;
-        memcpy(cursor + 4, pps[i].GetData(), pps[i].GetDataSize());
-        cursor += pps[i].GetDataSize() + 4;
-      }
-      return true;
-    }
-    return false;
-  }
-
-  virtual void UpdatePPSId(AP4_DataBuffer const& buffer) override
-  {
-    if (!needSliceInfo)
-      return;
-
-    //Search the Slice header NALU
-    const AP4_UI08* data(buffer.GetData());
-    unsigned int data_size(buffer.GetDataSize());
-    for (; data_size;)
-    {
-      // sanity check
-      if (data_size < naluLengthSize)
-        break;
-
-      // get the next NAL unit
-      AP4_UI32 nalu_size;
-      switch (naluLengthSize)
-      {
-        case 1:
-          nalu_size = *data++;
-          data_size--;
-          break;
-        case 2:
-          nalu_size = AP4_BytesToInt16BE(data);
-          data += 2;
-          data_size -= 2;
-          break;
-        case 4:
-          nalu_size = AP4_BytesToInt32BE(data);
-          data += 4;
-          data_size -= 4;
-          break;
-        default:
-          data_size = 0;
-          nalu_size = 1;
-          break;
-      }
-      if (nalu_size > data_size)
-        break;
-
-      // Stop further NALU processing
-      if (countPictureSetIds < 2)
-        needSliceInfo = false;
-
-      unsigned int nal_unit_type = *data & 0x1F;
-
-      if (
-          //nal_unit_type == AP4_AVC_NAL_UNIT_TYPE_CODED_SLICE_OF_NON_IDR_PICTURE ||
-          nal_unit_type == AP4_AVC_NAL_UNIT_TYPE_CODED_SLICE_OF_IDR_PICTURE //||
-          //nal_unit_type == AP4_AVC_NAL_UNIT_TYPE_CODED_SLICE_DATA_PARTITION_A ||
-          //nal_unit_type == AP4_AVC_NAL_UNIT_TYPE_CODED_SLICE_DATA_PARTITION_B ||
-          //nal_unit_type == AP4_AVC_NAL_UNIT_TYPE_CODED_SLICE_DATA_PARTITION_C
-      )
-      {
-
-        AP4_DataBuffer unescaped(data, data_size);
-        AP4_NalParser::Unescape(unescaped);
-        AP4_BitReader bits(unescaped.GetData(), unescaped.GetDataSize());
-
-        bits.SkipBits(8); // NAL Unit Type
-
-        AP4_AvcFrameParser::ReadGolomb(bits); // first_mb_in_slice
-        AP4_AvcFrameParser::ReadGolomb(bits); // slice_type
-        pictureId = AP4_AvcFrameParser::ReadGolomb(bits); //picture_set_id
-      }
-      // move to the next NAL unit
-      data += nalu_size;
-      data_size -= nalu_size;
-    }
-  }
-
-  virtual bool GetInformation(kodi::addon::InputstreamInfo& info) override
-  {
-    if (pictureId == pictureIdPrev)
-      return false;
-    pictureIdPrev = pictureId;
-
-    if (AP4_AvcSampleDescription* avc =
-            AP4_DYNAMIC_CAST(AP4_AvcSampleDescription, sample_description))
-    {
-      AP4_Array<AP4_DataBuffer>& ppsList(avc->GetPictureParameters());
-      AP4_AvcPictureParameterSet pps;
-      for (unsigned int i(0); i < ppsList.ItemCount(); ++i)
-      {
-        AP4_AvcFrameParser fp;
-        if (AP4_SUCCEEDED(fp.ParsePPS(ppsList[i].GetData(),
-                                      ppsList[i].GetDataSize(), pps)) &&
-            pps.pic_parameter_set_id == pictureId)
-        {
-          AP4_Array<AP4_DataBuffer>& spsList = avc->GetSequenceParameters();
-          AP4_AvcSequenceParameterSet sps;
-          for (unsigned int i(0); i < spsList.ItemCount(); ++i)
-          {
-            if (AP4_SUCCEEDED(fp.ParseSPS(spsList[i].GetData(),
-                                          spsList[i].GetDataSize(), sps)) &&
-                sps.seq_parameter_set_id == pps.seq_parameter_set_id)
-            {
-              unsigned int width = info.GetWidth();
-              unsigned int height = info.GetHeight();
-              unsigned int fps_ticks = info.GetFpsRate();
-              unsigned int fps_scale = info.GetFpsScale();
-              float aspect = info.GetAspect();
-              bool ret = sps.GetInfo(width, height);
-              ret = sps.GetVUIInfo(fps_ticks, fps_scale, aspect) || ret;
-              if (ret)
-              {
-                info.SetWidth(width);
-                info.SetHeight(height);
-                info.SetFpsRate(fps_ticks);
-                info.SetFpsScale(fps_scale);
-                info.SetAspect(aspect);
-              }
-              return ret;
-            }
-          }
-          break;
-        }
-      }
-    }
-    return false;
-  };
-
-  virtual STREAMCODEC_PROFILE GetProfile() override { return codecProfile; };
-
-private:
-  unsigned int countPictureSetIds;
-  STREAMCODEC_PROFILE codecProfile;
-  bool needSliceInfo;
-};
-
-/***********************   HEVC   ************************/
-
-class ATTR_DLL_LOCAL HEVCCodecHandler : public CodecHandler
-{
-public:
-  HEVCCodecHandler(AP4_SampleDescription* sd) : CodecHandler(sd)
-  {
-    if (AP4_HevcSampleDescription* hevc =
-            AP4_DYNAMIC_CAST(AP4_HevcSampleDescription, sample_description))
-    {
-      extra_data.SetData(hevc->GetRawBytes().GetData(), hevc->GetRawBytes().GetDataSize());
-      naluLengthSize = hevc->GetNaluLengthSize();
-    }
-  }
-
-  virtual bool ExtraDataToAnnexB() override
-  {
-    if (AP4_HevcSampleDescription* hevc =
-            AP4_DYNAMIC_CAST(AP4_HevcSampleDescription, sample_description))
-    {
-      const AP4_Array<AP4_HvccAtom::Sequence>& sequences = hevc->GetSequences();
-
-      if (!sequences.ItemCount())
-      {
-        kodi::Log(ADDON_LOG_WARNING, "No available sequences for HEVC codec extra data");
-        return false;
-      }
-
-      //calculate the size for annexb
-      size_t sz(0);
-      for (const AP4_HvccAtom::Sequence *b(&sequences[0]), *e(&sequences[sequences.ItemCount()]);
-           b != e; ++b)
-        for (const AP4_DataBuffer *bn(&b->m_Nalus[0]), *en(&b->m_Nalus[b->m_Nalus.ItemCount()]);
-             bn != en; ++bn)
-          sz += (4 + bn->GetDataSize());
-
-      extra_data.SetDataSize(sz);
-      uint8_t* cursor(extra_data.UseData());
-
-      for (const AP4_HvccAtom::Sequence *b(&sequences[0]), *e(&sequences[sequences.ItemCount()]);
-           b != e; ++b)
-        for (const AP4_DataBuffer *bn(&b->m_Nalus[0]), *en(&b->m_Nalus[b->m_Nalus.ItemCount()]);
-             bn != en; ++bn)
-        {
-          cursor[0] = cursor[1] = cursor[2] = 0;
-          cursor[3] = 1;
-          memcpy(cursor + 4, bn->GetData(), bn->GetDataSize());
-          cursor += bn->GetDataSize() + 4;
-        }
-      kodi::Log(ADDON_LOG_DEBUG, "Converted %lu bytes HEVC codec extradata",
-                extra_data.GetDataSize());
-      return true;
-    }
-    kodi::Log(ADDON_LOG_WARNING, "No HevcSampleDescription - annexb extradata not available");
-    return false;
-  }
-
-  virtual bool GetInformation(kodi::addon::InputstreamInfo& info) override
-  {
-    if (!info.GetFpsRate())
-    {
-      if (AP4_HevcSampleDescription* hevc =
-              AP4_DYNAMIC_CAST(AP4_HevcSampleDescription, sample_description))
-      {
-        bool ret = false;
-        if (hevc->GetAverageFrameRate())
-        {
-          info.SetFpsRate(hevc->GetAverageFrameRate());
-          info.SetFpsScale(256);
-          ret = true;
-        }
-        else if (hevc->GetConstantFrameRate())
-        {
-          info.SetFpsRate(hevc->GetConstantFrameRate());
-          info.SetFpsScale(256);
-          ret = true;
-        }
-        return ret;
-      }
-    }
-    return false;
-  }
-};
-
-/***********************   MPEG   ************************/
-
-class ATTR_DLL_LOCAL MPEGCodecHandler : public CodecHandler
-{
-public:
-  MPEGCodecHandler(AP4_SampleDescription* sd) : CodecHandler(sd)
-  {
-    if (AP4_MpegSampleDescription* aac =
-            AP4_DYNAMIC_CAST(AP4_MpegSampleDescription, sample_description))
-      extra_data.SetData(aac->GetDecoderInfo().GetData(), aac->GetDecoderInfo().GetDataSize());
-  }
-};
-
-/***********************   VP9   ************************/
-
-class ATTR_DLL_LOCAL VP9CodecHandler : public CodecHandler
-{
-public:
-  VP9CodecHandler(AP4_SampleDescription* sd) : CodecHandler(sd)
-  {
-    if (AP4_Atom* atom = sample_description->GetDetails().GetChild(AP4_ATOM_TYPE_VPCC, 0))
-    {
-      AP4_VpccAtom* vpcc(AP4_DYNAMIC_CAST(AP4_VpccAtom, atom));
-      if (vpcc)
-        extra_data.SetData(vpcc->GetData().GetData(), vpcc->GetData().GetDataSize());
-    }
-  }
-};
-
-/***********************   TTML   ************************/
-
-class ATTR_DLL_LOCAL TTMLCodecHandler : public CodecHandler
-{
-public:
-  TTMLCodecHandler(AP4_SampleDescription* sd) : CodecHandler(sd), m_ptsOffset(0){};
-
-  virtual bool Transform(AP4_UI64 pts,
-                         AP4_UI32 duration,
-                         AP4_DataBuffer& buf,
-                         AP4_UI64 timescale) override
-  {
-    return m_ttml.Parse(buf.GetData(), buf.GetDataSize(), timescale, m_ptsOffset);
-  }
-
-  virtual bool ReadNextSample(AP4_Sample& sample, AP4_DataBuffer& buf) override
-  {
-    uint64_t pts;
-    uint32_t dur;
-
-    if (m_ttml.Prepare(pts, dur))
-    {
-      buf.SetData(static_cast<const AP4_Byte*>(m_ttml.GetData()), m_ttml.GetDataSize());
-      sample.SetDts(pts);
-      sample.SetCtsDelta(0);
-      sample.SetDuration(dur);
-      return true;
-    }
-    else
-      buf.SetDataSize(0);
-    return false;
-  }
-
-  virtual void SetPTSOffset(AP4_UI64 offset) override { m_ptsOffset = offset; };
-
-  virtual bool TimeSeek(AP4_UI64 seekPos) override { return m_ttml.TimeSeek(seekPos); };
-
-  virtual void Reset() override { m_ttml.Reset(); }
-
-
-private:
-  TTML2SRT m_ttml;
-  AP4_UI64 m_ptsOffset;
-};
-
-/***********************   WebVTT   ************************/
-
-class ATTR_DLL_LOCAL WebVTTCodecHandler : public CodecHandler
-{
-public:
-  WebVTTCodecHandler(AP4_SampleDescription* sd, bool asFile) : CodecHandler(sd), m_ptsOffset(0)
-  {
-    // Note: the extra data must be exactly of 4 characters as imposed in kodi core
-    if (asFile)
-    {
-      // Inform Kodi subtitle parser that we process the data as single file
-      extra_data.SetData(reinterpret_cast<const AP4_Byte*>(&m_extraDataFILE), 4);
-    }
-    else if (sd) // WebVTT ISOBMFF format type (ISO/IEC 14496-30:2014)
-    {
-      // Inform Kodi subtitle parser that we process data as ISOBMFF format type
-      extra_data.SetData(reinterpret_cast<const AP4_Byte*>(&m_extraDataFMP4), 4);
-    }
-  };
-
-  virtual bool Transform(AP4_UI64 pts,
-                         AP4_UI32 duration,
-                         AP4_DataBuffer& buf,
-                         AP4_UI64 timescale) override
-  {
-    m_data.SetData(buf.GetData(), buf.GetDataSize());
-    m_pts = pts;
-    m_duration = duration;
-    return true;
-  }
-
-  virtual bool ReadNextSample(AP4_Sample& sample, AP4_DataBuffer& buf) override
-  {
-    if (m_data.GetDataSize() > 0)
-    {
-      buf.SetData(m_data.GetData(), m_data.GetDataSize());
-      sample.SetDts(m_pts);
-      sample.SetCtsDelta(0);
-      sample.SetDuration(m_duration);
-      // Reset the source data size otherwise we send always same data without end
-      m_data.SetDataSize(0);
-      return true;
-    }
-    buf.SetDataSize(0);
-    return false;
-  }
-
-  virtual void SetPTSOffset(AP4_UI64 offset) override { m_ptsOffset = offset; }
-
-  virtual bool TimeSeek(AP4_UI64 seekPos) override { return true; }
-
-  virtual void Reset() override {}
-
-private:
-  AP4_UI64 m_ptsOffset;
-  AP4_DataBuffer m_data;
-  AP4_UI64 m_pts{0};
-  AP4_UI32 m_duration{0};
-  const AP4_Byte m_extraDataFILE[4] = {'f', 'i', 'l', 'e'};
-  const AP4_Byte m_extraDataFMP4[4] = {'f', 'm', 'p', '4'};
-};
 
 /*******************************************************
 |   SampleReader
